several updates
1. added 'Miss!' indicator when attacks do not pass hit calculation. i may add a separate 'dodge' indicator to indicate a miss while a target is in a dodge animation (would only apply for the Player, and potentially enemies/bosses that actively dodge the Player's attacks).
2. added music to the Snow scene. music is not final as i have a couple other 'snow' tracks to potentially use.
3. added alpha fades to XP popups before despawning, and a quick light glint on the text when it pops to make it easier to notice.
issue:
the text popups seem to be popping up to the right of where they should be. i'd like them to be centre-aligned with the global_position of the hitboxes that spawn them, but they aren't. i have tried adjusting the transform position.x value of the damagePopup Node and its label child; i have tried changing the text-align of the label; i have tried changing the horizontal growth of the text box, and nothing seems to have worked. actually, horizontal growth works - it grows in the right direction (left), but it still originates to the right of where it should be, and i can't identify exactly why so i don't really know how to go about properly adjusting it.