working on backstep and queuing actions from it
big update for the Backstep action:
first, backsteps can now be chained consecutively. because of this. i have applied friction just before the backstep ends in order to give it some "weight". also, when the friction is applied, base_enemy_accuracy is set to 50% (from 16% during a backstep; default is 66%)
the input logic during a backstep is a bit complex:
first, it checks if the "attack" action is RELEASED, as this is how charged attacks are unleashed. queues the appropriate charged attack for when the animation finishes; if no attacks are charged, queues a regular attack.
if the release of the "attack" action is not detected, it will monitor if the "roll" button is "just pressed", meaning it will have to have been pressed DURING the animation. if the player has stamina, queues a backstep.
next is the backstep_stop method touched on above. sets backstep_moving to false, which is monitored during the backstep_state in order to apply the appropriate velocity/friction to the player. also sets base_enemy_accuracy to 50.
finally, is the backstep_animation_finished method.
if an action was queued during the animation, this is where it will start. some attacks queued from a backstep have additional properties. for example, the shade attack has forward motion applied to it (75% of roll_speed); the flash attack has forward motion as well (100% of roll_speed), in additino to setting base_enemy_accuracy to 0... basically, during a queued Flash attack, you are very, very difficult to hit.
if none of that happens, there is a check to see if the the "attack" button is pressed/held. if it is, the backstep_animation_finished is treated as an attack_animation_finished, and goes directly into the charge_state from a backstep - no attack needed. this is handy for starting a charge as quick & efficiently as possible, as backsteps are both less expensive stamina-wise & time-wise than a regular attack.
if a regular attack is queued (as in, during the backstep, it was detected that the attack button was released), adds velocity (75% of roll speed) to the player when the backstep completes, effectively resulting in a forward-dash attack (the same as backstep attacks in dark souls).
the final conditional is if none of the above occured: treats the end of the backstep animation as the end of an attack animation. the reason why the regular queued attack condition had to be sandwiched in between the "charge after backstep" and "do nothing after backstep" is because of the velocity modifier to the attack, and the check as to whether the attack button is held needs to occur first.
it is probably overly complicated and badly coded in some parts, but it seems to work properly. needs a lot of testing and tuning; i'd noticed bugs where functions that run during an animation run multiple times if the state that plays the animation is re-engaged without that state changing first. this caused problems with chaining backsteps together, as well as detecting whether queued attacks were input. i'm not sure if this bug is completely gone yet, but it seems to work fine so far.
overall i'm happy to have made it this far as it feels i'm doing some real game design at the moment.