increased regular attack speed
it's 10% faster than it was before which feels a lot better, but i need to consider balancing the feeling of character growth with the feeling of tight control. should the speed stat effect attack speed? how much?
this is also a programming consideration as the animations have fixed lengths. i have included a way to change the timescale of the animations, as that is the only way to change animation speed through code at the moment. the issue with this is that it will be difficult to minutely tune values exactly. i will try experimenting with this, but first a brainstorm:
the player's default stats all start at 4. let's say at the "endgame", stats will be around 40. first, we need to come up with what will be the ultimate difference between the slowest default attack speed and the maximum default attack speed. i am going to use 20% as a ballpark since it's mathematically convenient.
so, our goal is to get from 1.0 at default stats to 1.2 with high stats.
if each point in SPD increases the timescale by 0.005, then 40 points in SPD would increase it by 0.2 (20%) to 1.2. this looks fine on paper but will have to see how it feels in the game.
the basic attack animations have just had their length reduced by about 25%, which feels much faster than before. speeding it up further by 20% over the course of the game shouldn't be a gigantic gap; it will likely go barely noticed, but if the animations are compared side-by-side, it will always be clearly obvious which is faster. individual differences between SPD increments (.5% attack speed increase) will be barely noticeable between levels, but the effect will be felt over time, which is the goal.