critical hits implemented / working on pixel shaders
crits need testing and such
i've been trying to find a way to render the font outlines pixel-perfect, and turns out, it's fucking impossible. shaders i've looked at online by others are ugly as fuck and or glitchy, and the one pixel-perfect shader i've found works on sprites but not text. i found a global pixel perfect renderer which sidesteps the problem and renders everything as is pixel-perfect, but this causes visual glitches in other areas so i won't bother with it for this game (maybe for next game, if it's built from the ground up to be pixel-perfect, it would be more useful)
if i can't fix this, i might have to learn how to code a shader myself, but honestly if it can't work on sprites i'm not sure what the deal with text is. i'm 100% not the only person who's encountered this problem, so somebody out there must have a simple solution, because i don't think it's that complex an issue.