adjusting Crow collision shapes / crit_rate effected by dexterity_mod
the crow's hurtbox & hitbox now adjust for when it is in flight, making it much easier to judge both whether you'll hit it or not, as well as whether it'll hit you or not.
the crit rate of attacks are now effected by the player's dexterity modifier, applied for special attacks. i just need to figure out how MUCH the dexterity_mod should effect crit chance. currently, it effects crit chance twice as much as base dexterity does.
example in practice:
base_crit_rate: 4%
dexterity bonus: 1% (0.25% per point in DEX)
dexterity modifiers: 0 for regular attacks, 4 for Lv.1, and 8 for Lv.2. making the dexterity_mod bonus 2% for Lv.1, & 4% for Lv.2.
final_crit_chance:
5% for regular attack
7% for Lv.1
9% for Lv.2
so, basically, crit chance scaled much higher when an attack has a dex modifier. this will most likely have to be adjusted to balance things throughout the game, but these numbers seem fine to work with for now.