applying pixel boundary text shader to labels
found a simple pixel shader that works on text for pixel art, and have applied it to all text popups that occur (damage, level up, etc.). not only does it make the outline sharp, it looks better when text fades as well.
modified the Level Up display. it's bigger to make it more noticeable, and its position is locked to the player's. it looks much nicer this way. i don't intend to keep the levelling system as is, though, so it will likely change in the future.
also adjusted the positions of various elements, hoping it's cleaner.