refactored crit popup into damage popup; got rid of it
i'm a real programmer now!
the next big programming task i need to do is come up with a generic "enemy" class, which might be problematic because of the differences between the bat & crow behaviour. i have to come up with a way to standardize the data used by both the bat and crow (eg., displaying damage values when hit; hurtbox behaviour) into something common. once that is done, i will be able to use it for every enemy going forward, saving a lot of time and potential headaches in the future.