working on timed death
the code is pretty messy. i expected this way of doing it wasn't going to work exactly the way i wanted and was anticipating re-programming everything to take into account all the things i decided i wanted to modify, but i managed to get everything working so all i have to do now is hope nothing breaks. if i want to change things further, or if there's bugs i haven't seen, i will re-do it.
here's how it works:
if an attack would bring you down to 0 HP, instead of dying right away, you lose all stamina and become slowed * sweating. during this time, you have a chance to drink a potion to prevent yourself from dying, however the potion will be 1/3rd as effective, so you are incentivized to heal BEFORE getting dealt a deathblow. this is meant as a "last chance" kind of thing - i want the game to be a fun challenge, but not overly frustrating and punishing... only slightly.