refactoring chapter display script / working on scene transition
currently, selecting New Game on the title screen brings you to the control screen.
pressing start on the controls screen instances the "chapter display" scene, which runs on a timer. when it times out, changes the scene completely to the Home scene.
instead of the controls screen instancing the Chapter Display as a child of itself, the Chapter Display should be separate/discrete from the rest of the program and able to run on its own, facilitating smoothly fading/transitioning into gameplay. the textbox ("Chapter One") string should be exported, allowing the scene to be reused.