beginning work on stat allocation box
this may or may not be extremely complicated
i would like to make this as discrete/separated from the generic dialog box as possible.
the container will be its own scene, instantiated by the dialogBox when it recieves a "level_up" event.
the button/control/signal programming will be hard coded into the levelUp container, ie. there shouldn't be any parsing of dialog events to create & connect buttons.
the level up dialog box might be better off being instantiated COMPLETELY from scratch with NOTHING to do with the current DialogBox.tscn
if things are coded into the LevelBox itself, in theory all the DialogBox needs to do is recieve a "level_up" event, which will then instantiate the LevelUp_Container with everything hard-coded already. the player picks and selects a stat to level up as if they are picking a dialog option, that stat gets updated, and the DialogBox behaviour carries on as usual.