working on stat allocation interface
it's sloppy, but -
i've duplicated the DialogBox code for the DialogLevelBox because i'm not sure how much more of the default DialogBox's functionality i'm going to want to use in the future regarding the LevelBox, and at the same time i want the LevelBox behaviour to be totally separate from DialogBox behaviour. i can re-factor and clean up redundancy/useless code later.
the dialogLevelBox does its thing when it recieves a "level_up" event in its "dialog_script": it instantiates a LevelUp_Container (a scene containing all the stat buttons prelabelled), then individually connects each button in the LevelUp_Container to a function in the LevelBox called _on_level_selected.
the LevelBox has a variable called stats_remaining, which defaults to 2. this variable refers to the number of allocatable stats. when a stat button is pressed, ie. when the _on_level_selected function is run, this variable decrements by 1. the idea is to have this variable change over time: you begin with the ability to allocate 2 stats per level, but maybe that can change later.
once stats have been allocated, deletes the LevelUp_Container and runs the load_dialog() function (basically identical to what happens when you select a dialog choice in the default DialogBox). since there is nothing after the "level_up" event in the LevelBox's dialog_script, it closes, and is deleted.
the reason i want to load_dialog is because i am going to include an "are you sure?" confirmation for stat allocation; it's not in place yet, but the idea is to load_dialog to the "are you sure" question, then a yes/no option. if yes, the entire LevelBox is deleted and the selected stats are added to the PlayerStats. if no, the plan is to have stats_remaining and the relevent stats_to_add (?) revert to their default values and the level up process begins anew.