stat allocation... kind of working
what a mess. choosing to work with/from the original dialog box has been both a blessing and a curse. i guess on the upside it has revealed a number of potential flaws/vulnerabilities of the dialog box.
the primary issue seems to be with handling inputs. it's very difficult to direct exactly how an input event propagates through the program. it's almost impossible - it has a fixed order of operation that is not the most intuitive.
it's not implemented into gameplay yet because gaining multiple at once is not working.