bugfix and addition
game no longer crashes when zoning while the level up display is on screen.
also, added a "fade" when zoning. unfortunately, this has introduced an issue where you can see the "default" player position for a brief moment when zoning to a new scene, as opposed to the position provided by the Exit node. i'm not sure why this happens when the fade occurs entirely in the previous scene. my guess is that the new scene gets loaded/is run too fast before the player's new position is set. i don't know why this happens so i will look into it.