working on scene fades
transition between forest/home getting smoother.
audio sounds OK.
the FadeIn scene is instantiated automatically when a World is created. a dirty way to override the Chapter Display is to simply instantly delete the FadeIn if global.chapter_display = true.
right now the fades are a solid 1 second, which feels a bit long to me. this will be adjusted. it'd be cool to have it customizable for easy longer/shorter fades for more or less dramatic/cinematic effects.
2 issues:
1. zoning while the Level Up message is onscreen will prevent its position from being updated in the new scene. my theoretical solution is to have the Level Up notice NOT instantiated as a child of the Player - which makes its position update automatically - but to have its position set by the RemoteTransform2D child of the Player, which is how the camera moves/follows the player. i'm not sure why this happens honestly because the new Player (the one created upon zoning) should be a new instance, which means its previously instantiated children (ie. the level notice) shouldn't be in the new zone... unless i'm missing something in my own code, which is likely.
2. this is something i've ignored for a long time, but the player's direction vector should be stored and sent to new instances of the Player. i'll probably just add it to the PlayerStats and have it get updated automatically. either that, or manually setting it on each Exit, like the position, but that would be more manual work. would probably end up running better, though.