modifying crow hit effects - looks/feels better
half the time, a HitEffect's sprite is flipped horizontally, providing some easy variance.
the amount of damage of a hit affects the range within which its HitEffects (feathers) can spawn, as well as the amount of feathers (HitEffects).
feathers get deleted when off screen.
in addition to the crow's normal death animation, 4 additional HitEffects get spawned, adding some random variance to crow deaths.
i would like to figure out a way to adjust the randomness of the range at which feathers spawn. right now, they can spawn a maximum of 32 pixels on both axes away from the crow's global_position (if less than 32 dmg is dealt, then that is the maximum range used, in pixels). if something does, say, 80 damage, that means 10 feathers will spawn. my issue is that the randomness applied to the range is an equal distribution - in other words, i would prefer it if values closer to the extremities were less common and values closer to the origin were more common.