mitigating previous issue
first, i halve the damage_count value if damage dealt is above 80, effectively making the max 40. then 4 is subtracted from this value for each Feather spawned, for a maximum of 10 feathers. this means that the potential spawn range of each individual HitEffect varies more with each tick of damage_count.
for example...
80 damage is dealt.
damage count is damage/2 OR 40 - whichever is less. so, 40.
the area in which the first feather spawns is based on the damage_count, so -40 to 40, or a total area of 80x80.
4 is then subtracted from damage_count, making it 36, therefore the effective spawn range of the feathers becomes 72x72. the next feather's effective spawn range area is 66x66, and so on.
this seems to make sense for now but will likely need to be adjusted depending on damage values. at default stats, it means 1 or 2 feathers spawn on hit, maybe 2-4 on a crit, and up to 10 on landing a special attack. when strength is increased, you can get 10 feathers to spawn on each hit.
one potential issue - and this is only for feathers that spawn far away - is their position is set instantly right when they spawn, meaning sometimes it can look as if they appear out of no where. it might be complicated to do (as i'll need to write a function using the delta argument), but it might be cooler/more satisfying if the hit feathers spawn at the hit's origin point and quickly move outward.
there were some unrelated issues i've noticed: zoning when at less than 100% HP will have the health bar backround gradually empty to the current health upon reinstantiation. this shouldn't happen.
also, i've noticed sometimes the charge level doesn't get incremented properly, ie. letting go of the attack button while at Lv.2 results in a Lv.1 attack. not sure why this happens.