feathers move from origin point to target
when a hit effect is created, its global_position is set to that of the hurtbox that created it. then some random values (based on the amount of damage done) are combined with the global_position to create a target_position. the feather moves from its origin point to the target_position before stopping.
this happens very fast, so the effect is barely noticeable, but it looks better than before.