overly complex issue fix (probably)
noticed an issue where if you zoned while in the "dying" state, you would spawn out of the dying state, but at 0 HP. the previous issue of the "health drain" resetting still happens, and when it reaches 0 (when dying is supposed to become dead), the game crashes.
my solution is to prevent players from zoning while dying. how to do this? add a PhysicsBody2D to the Exit's Area2D, and give it the same collision shape. the Area2D is on collision layer 12 "Exits", and the PhysicisBody2D shape is on collision layer 13 "ExitBlock". basically, when the player becomes "dying", their layer 13 collision mask is set to true, meaning they can interact with the PhysicsBody2D shape of the exit, meaning the Player's body gets blocked from ENTERING the Area2D's shape if they are "dying". simply put, the player's PhysicsBody is blocked by the Exit's PhysicsBody, but only while dying. as an additional measure, the Exit's Area2D checks the player's dying variable and returns if true, preventing the rest of the Exit code from running.