balancing / attacks / various adjustments
quadrupled charge time for testing purposes. this makes me realize i can tie charge time to a stat like speed, as it is simply a hard-coded value for now.
shade lv.2 attack has nerfed range, which makes you open for longer when the attack ends.
crow eyes glow in the dark.
added blood to bats.
fixed a bug where sometimes a Lv.1 charge attack would come out with a Lv.2 charge. it still happens sometimes but... less. there seems to be issues when the player is charging while zoning, as well as if there are obtainable inventory items around, i think. not sure what is preventing the Lv.2 from coming out.