adjusting visual elements / re-factoring player reference code
fixing GUI and other adjustments
as an experiment, i may try to instantiate a second Player to see how viable 2P would be to implement at this time. i'm not sure if this means i have to find every single reference to the Player in code and duplicating them, though.
i've managed to refactor most of the hard-coded references to the player in the Crow scene. hard-coded elements include checking the player's health to determine health drops, applying experience points to the player, and calculating distance to the player to determine whether to de-spawn or not.
the first two elements are easily duplicatable. i could even probably write a Global experience distribution function to handle that, as well as a Global function to check both [theoretical] players' health values.
i also have an idea to replace the distance check. instead of checking distance to the player, it will check if it's off-screen. simple enough.
one thing that will be difficult to implement is camera control and limiting the players from moving off-screen. games like secret of mana seem to have a "wall" or physical collision barrier around the viewport. otherwise, the camera's position could possibly be set by the average of both players' positions.
i'll attempt this as a challenge for today but i'll move on if i can't get it working.