created player 2 / new GUI
dirty implementation, but working so far.
basically copied the Player class and got rid of the code that creates a player inventory for player2 (inventory will be shared). added new controls and a new stats scene. world scene checks if "multiplayer_2" is enabled; instantiates Player2 if true, and sets its position to that of the default Player.
Item handling is dirty. use_item() now takes an argument, which is an integer to determine who it was that used the item. so use_item(1) uses the potion on player 1, and use_item(2) uses the potion on player 2.
experience distribution function created in Global script. when run, checks the multiplayer_2 variable. if true, divides the exp earned and gives it to the players.
issue:
on levelling up, the level up displays for both players appear on player 1. i believe i'm going to remove the "following" level up notice and just have it be a static notice tied to the UI. this should make fixing the problem easier, if it doesn't simply fix the problem outright.
a number of things have yet to be implemented:
hit & crit calculations only take into account Player 1's stats, because the reference to PlayerStats has been hard-coded into the algorithms.
potential solution:
apply relevant Stat parameters to the hitbox for the hurtbox to take from, as opposed to the hurtbox grabbing them directly from PlayerStats.
so... for hit calculation, this would be:
base_accuracy, dexterity, & dexterity_mod
& crit calculation:
base_crit_rate, dexterity & dexterity_mod
damage is already sent from the hitbox, so that's fine. just need to add the others and make sure they update properly with PlayerStats.
another big issue has to do with how enemies spawn and despawn. if enemies are far away enough from player1 to despawn but close enough to player2 to spawn, then a ridiculous amount of "spawners" get created, usually freezing the game.
the final and biggest issue is how to handle the camera. maybe i will make a Global function that averages the positions of both players, applies value to the camera's position. this won't prevent players from going off-screen, though. have to figure out a good solution for that. once i figure out camera control, i will re-program from scratch how enemies spawn/despawn using viewport visibility as opposed to collision boxes.