figured out 2-player camera control
if i could put a collision box around the viewport, that'd be good.
as of right now, during each player's move function there are 4 checks. each check is for a particular distance in a particular direction, and if a check is passed, it moves (transforms) the OTHER player, creating a "drag" effect when you walk to the edge of the screen and the other player isn't moving.
the camera settings needed LOTS of adjustments and tuning to get it where it is, so i've saved an instanceable version:
new Camera2D scene for default camera settings for multiplayer and have put it in the System folder. while typing that, though, i realize that the settings may be different depending on whether you're playing with 1 or 2 people, because camera behaviour is VERY different. so i eventually will probably have to set default settings via script depending on whether multi or single player.
the camera position is the average of both players' positions.
i've also changed other various things around. added a short delay audio bus for the new crit sound, and moving the level up notices to be anchored to the GUI.