level up stat selection functioning properly
behaves fine when multiple level ups happen. haven't programmed it for player 2 yet and probably won't prioritize that for now.
really need to work on the visuals for the level screen. gene audio as placeholder for now
i took away the confirmation for stat allocation. picked the wrong stat? fuck you! though it only confirms when you let go of the button. if you move the cursor before letting go, you won't allocate.