working on level up / stat allocation menu / display
figuring out buttons is infuriating.
there are a number of issues. one was the game became unpaused during a level up. buttons lost focus, stats were not allocated, and the selection remained on screen. have not been able to recreate.
a
HP values are displayed above the health bar.
player2 UI is deleted if game is started in single player mode (GameManager.multiplayer_2 = false)