various adjustments
the staminadisplay in the ChargeUI now has a short fade-in and fade-out when it displays/hides.
time is slowed by 20% while in the 'dying' state. i basically lowered the timescale of the engine to 80% while dying, so it's probably not the safest way to do this. i just wanted to see the effect.
fixed a bug where experience would not be awarded if the enemy dies off-screen. now waits until the enemy completes its dying state before deleting itself.
a major bug can happen during level ups. i believe what is happening is that the DialogLevelBox interprets an input as the input to END/close the dialog, before any stats gets chosen. the greyscale and stat buttons remain on screen, and stats don't get allocated. not obvious what is causing this issue, will have to do rigourous testing to localize the issue. i have a feeling it occurs when level ups are consecutive, as there are a lot of timing events involved with that. i have a feeling an input is interpreted wrong during one of these timing events.