troubleshooting consecutive level ups
the issue consistently appears when 3 or more level ups happen at once. two seem to be fine, though occasionally (>10% of the time) they're not.
i think i've localized the issue to be something wrong in the Player's level_up() method. when 3 or more level ups occur simultaneously, it appears that the first level_up() queues TWO (or more) level up dialog boxes at the same time - and only one listens for inputs, so when stats are chosen, it is deleted and the game resumes while the other one stays perpetually on screen.
the levelTimer uses yield which i think is a mistake as i believe two or more parellel processing lines get created. thus the two level_ups instantiated at the same time both believe just_leveled is 0, so it spawns two level boxes.