new level_up() function working nicely!
well, that was fast. proof that it's always better to conceptualize what EXACTLY you want the computer to do before attempting to actualize what you want through trial & error coding.
now have to work on the level up screen a bit. a counter to show the amount of stats remaining, as well as a better overlay of current stats and stats after level. some descriptive text to pop up when a stat is highlighted to explain what the stat does will also be necessary.