various fixes & adjustments
killing blows do not send enemies flying as far (10% less).
the 'Patience' button text displays properly on pressed now (did not light up like the others).
when Willpower (vitality) is increased, the amount of max_health added is also added to the current health. in theory, this should allow a dying player to be saved by levelling up if at least one point is put into Willpower. works, but hasn't been thoroughly tested yet.
the backstep animation now has additional spriteframes depending on input_vector.
the shade animation has been tweaked slightly.
you are now able to begin charging another attack right away after a shade attack if the button is held while the animation ends. if the attack button is detected to be released during the shade animation, attack2 is queued instead. the logic here is iffy and i'm not sure it's scaleable because of the overlap of attack and special attack functions. a dirty check in the attack_animation_finished() logic applies only to the shade state, as the shade attack is the only attack currently that can combo directly into attack2. the check in question is:
if attack2_queued && !Input.is_action_pressed("attack")
what this means effectively is that in order for attack2 to get queued, the attack button must NOT [!] be held down at the end of the previous animation. in other words, if your current attack can combo into attack2, the attack button must be released before the current attack ends. otherwise, you'll simply begin charging.