stamina regen levels implemented
while in the MOVE state, runs the stamina_regeneration() function, which is to where the previous stamina regeneration code has been moved.
the old hard-coded stamina_regen_rate now has its own variable.
stamina_regeneration() first checks if the player is sweating, and sets the stamina regeneration rate to 60% of the base value if true.
if not sweating, the stamina_regeneration() function matches the new stamina_regen_level variable with the appropriate regeneration value. currently, there are three levels:
0, 100% regen;
1, 200% regen;
and 2, 1000% regen.
if "attack" or "roll" are pressed, returns out of the conditional (preventing the timer that increments stamina_regen_level from starting while a button is held down).
then, checks 2 conditions for stamina_regen_level to be incremented: stamina_regen_level must be less than 2, and the timer must be stopped.
if both == true, starts a 1s timer, at the end of which stamina_regen_level is incremented by 1.

another function, stamina_regen_reset(), has been created. this function is called when the player finished a "getting hit" animation, and when the player finished an attack or roll animation. this checks the current stamina_regen_level and sets it to 0 if it's above 0. it also starts a timer, preventing the stamina_regen_level from being incremented until it times out.
i'm still trying to tune the timer values appropriately, as there is still some weirdness happening in the code. for example, when unleashing a shade attack, the stamina_regen_level gets incremented during the animation, and reset back to 0 at the end of the animation. no stamina regeneration occurs because the Player is not considered in the MOVE state, but the regen_level should not be incremented either way. my dirty workaround is to check if the timer is running if the stamina_regeneration() function (in the MOVE state) detects a stamina-consuming action button to be pressed, and to start the timer if it does, preventing the stamina_regen_level from incrementing until said timer times out, which is 0.4s. in practice, this timer gets started every 0.4s while charging in order to prevent the level incrementing, because it checks to see if the timer is_stopped().