stamina charge / damage value adjustments
adjusted stamina charge behaviour. there are now 5 'tiers' of regeneration, with each tier increasing the base stamina regen rate by a factor of 2.
at the end of every action that consumes stamina, a 1 second timer is started. the stamina regeneration tier advancement timers do not begin until this timer times out.
after 1 second, and if no attacks/rolls are performed, a 0.6 second timer starts. when this times out, the stamina regeneration rate becomes multiplied by 2, and another 0.6s timer is started, upon whose timeout multiplies the stamina regen rate by 4, then 8, and finally 16.
the idea is that if you you have a very high stamina cap, if you relent on attacking and focus strictly on movement for a few seconds, you can regain all of your stamina in a matter of seconds as opposed to waiting for it to regenerate at the base rate over time. the intention is not to take importance away from stamina management, but also to minimize stamina regeneration "downtime". obviously it will have to be balanced over time but the basic framework is in place.
also fixed a bug where the proper damage values were not being set properly for the player.
also, the "sweating" indicator appears OVER the stamina wheel now, indicating more clearly that you should relent from attacking/stop sweating before carrying on with combat actions.