damage popup text follows hit object / refactoring some code
having the damage pop follow the thing that was hit makes damage clearer.
potential issues:
the text display is not centered. it begins in the center, and writes to the right. potential non-issue, but not obvious to solve. currently, all hurtboxes (including the players' and enemies') share the same script. i was thinking of separating the player's and the enemies' hurtbox script into two to allow for divergent damage pop behaviour. for example, i might want the damagePopup to appear next to the enemyHealthBar. i also might want multiple damagePopups on the same enemy to be sorted so that they don't overlap. this would involve adding damagePopups as children of a VboxContainer attached to the object. might be complicated to do, but it will be worth it i think.
i've also refactored the blood effect to the Global script. i want to have as many re-usable scripts as possible available globally.