bugfixes & gameplay adjustments
the bug where if you leveled with a special attack charged and did not continue to hold the attack button after the level up screen goes away resulting in a "stocked charge" has been fixed:
when the dialogLevelBox runs its end_dialog() method, a Global method called reset_input_after_dialog() is run, which basically runs a Player function check_attack_input(), which checks if the attack button is NOT pressed, and if it isn't, stops the charge and resets the charge meter. maybe, before stopping/resetting the charge, it should check which charge level the player is at and do an appropriate attack - i can see this leading to multiple unintended special attacks unleashed following a level up, though.
i've also removed temporary invincibility on hit from enemies, as i'm not sure how useful it was.
i've updated the Frenzy status (applied by pressing 'T'). first off, attacks' stamina drain is reduced by 80%, and knockback applied to enemies is reduced by 90%. the result is the ability to pummel enemies with a flurry of blows. next, in order to make this more useful, i should add a stun feature to enemies, where if you hit them multiple times in a row they become "stunned" and unable to attack or move for a few seconds.