level up bugfix
fixed a bug where the level up screen would close before the player had a chance to choose stats... hopefully.
i identified the problem as the levelBox's input check failing the two conditions when a levelBox is spawning in. the first checks if the delayTimer is running, and discards the input if it is. the second checks if "waiting_for_input" is true (it should only be true when buttons are on the screen and one is focused). if these conditions fail, it closes the levelBox.
my guess is that if the levelBox recieves an input on the frame the Timer stops but before the buttons are created/focused (and waiting_for_input to be set to true), the input is interpreted to close the levelBox. so i added another parameter to the first conditional, making it easier: check if [not] !waiting_for_input. if EITHER the timer = is_stopped or !waiting_for_input, sets the input as handled WITHOUT closing the levelBox.
the issue wasn't easy to replicate, so more rigourous testing is required to make sure it never happens again.