beginning code for spellcasting Ingredient Pouch
i am using the base Inventory script as foundational code to determine basic Pouch behaviour. so the Pouch will be an array, and each index will reference an ingredient.
the reason i wanted to use the base Inventory code is because the basic functions work properly, although the code is more complex than it needs to be. Pouch resources (ingredients) do not need to have additional parameters associated with them, they should only have quantities. so the way the current Inventory resources and Item resources is perhaps overkill, but I'm afraid of running into issues if i were to do it from scratch. so, if i can get it to work this way, i will be able to write it from scratch without issues either way.