organization / Pouch prep
moving all item-based resource stuff to assets/Resources folder.
created Ingredient.gd script, from which individual ingredients will be created.
created two Ingredient resources: Rock & Clay.
ran into issues when trying to contain Ingredients values in the Inventory resource. for some reason, when the inventory's _ingredients array was created, its value was set to that of _items. i literally spent hours going over every line of item-related code multiple times to look for typos or mismatches references. there were none. at a loss and frustrated that i had wasted hours of time, i decided to create a separate resource for ingredients: Pouch, as well as a separate script for compiling IngredientResources: IngredientDatabase.
when moving all the previously ingredient-related code from the Inventory script to the Pouch script, the issues i was having magically disappeared. there was nothing in the code that should have set player.inventory._ingredients to whatever _items was. the methods had completely different names and variables, so i am absolutely dumbfounded as to why it didn't work, because the code is essentially the same as it is now, simply localized into one resource as opposed to two. now the project is messier for seemingly no reason, but it doesn't break/crash yet, which is more important i guess.
an issue encountered was how the player gets RE-initialized after loading in for the first time. when getting initialized, the player's inventory is set via file (user/inventory.tres). since there are now two resources to load, that means the "inventory" key of the "new_inventory" variable being passed through by the Exit script into the Global "attributes"/scene change function should carry both the player's previous inventory & pouch. the solution was to make "new_inventory" an array, and have the player's inventory and pouch resources exist in different indexes of the array.
when the player gets re-initialized, sets the inventory to argument 0 of the Global attributes' "inventory" key and the pouch to argument 1.